# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Api Docs",
    "version": "1.0"
  },
  "paths": {
    "/Gateways": {
      "get": {
        "tags": [
          "Gateways"
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/QueryObject"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Gateways"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/SaveGatewayDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveGatewayDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveGatewayDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SaveGatewayDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Gateways/{id}": {
      "delete": {
        "tags": [
          "Gateways"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Gateways"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Peripherals": {
      "get": {
        "tags": [
          "Peripherals"
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/QueryObject"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Peripherals"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/SavePeripheralDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavePeripheralDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SavePeripheralDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SavePeripheralDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Peripherals/{id}": {
      "delete": {
        "tags": [
          "Peripherals"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Peripherals"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "QueryObject": {
        "type": "object",
        "properties": {
          "skip": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "take": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "filter": {
            "type": "string",
            "nullable": true
          },
          "sort": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SaveGatewayDto": {
        "required": [
          "ipv4",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "format": "string"
          },
          "ipv4": {
            "pattern": "\\b(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\b",
            "type": "string",
            "format": "string"
          }
        },
        "additionalProperties": false
      },
      "SavePeripheralDto": {
        "required": [
          "dateCreation",
          "gatewayId",
          "status"
        ],
        "type": "object",
        "properties": {
          "vendor": {
            "type": "string",
            "nullable": true
          },
          "dateCreation": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "boolean"
          },
          "gatewayId": {
            "type": "string"
          }
        },
        "additionalProperties": false
      }
    }
  }
}